video_summarizer/requirements.txt
# ============================================
# CPU SYSTEM REQUIREMENTS (Current PC)
# ============================================
# For GPU system, use: requirements-gpu.txt

streamlit
openai-whisper
moviepy
opencv-python
yt-dlp
numpy
faster-whisper
